### 2019-03-29 #1353 @whedon
[PRE REVIEW]: WGS2NCBI - Toolkit for preparing genomes for submission to NCBI
**Submitting author:** @rvosa (<a href="http://orcid.org/0000-0001-9254-7318">Rutger Vos</a>)
**Repository:** <a href="https://github.com/naturalis/wgs2ncbi" target ="_blank">https://github.com/naturalis/wgs2ncbi</a>
**Version:** v1.0.0
