### 2019-03-29 #1353 @danielskatz
@rvosa, please check your paper source, using https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography as an example/template.  When you've made changes, enter ```@whedon generate pdf``` as a new comment here, and iterate until you are happy with the paper.  Then we'll start the pre-review process by assigning an editor.
