### 2019-03-31 #1353 @pjotrp
> wave bioinformatics editing team (@pjotrp @mgymrek @brainstorm @lpantano) - can one of you take on this submission as editor?

We'll take it on. Thanks @rvosa for submitting this interesting tool!
