### 2019-04-03 #1353 @arfon
> @danielskatz @arfon I think whedon doesn't like me anymore ..... it doesn't assign me as editor, maybe I lost permissions :/

Hrm, not sure what's up here.
