### 2019-04-04 #1364 @whedon
[REVIEW]: WGS2NCBI - Toolkit for preparing genomes for submission to NCBI
**Submitting author:** @rvosa (<a href="http://orcid.org/0000-0001-9254-7318">Rutger Vos</a>)
**Repository:** <a href="https://github.com/naturalis/wgs2ncbi" target ="_blank">https://github.com/naturalis/wgs2ncbi</a>
**Version:** v1.0.3
