### 2019-04-11 #1364 @druvus
@rvosa I started to add issues to the wgs2ncbi repo regarding failing automatic tests, stalled help command and contribution documentation.  Could you please take a look at them?
