### 2019-04-11 cd7e0ec
tests now pass, added installation instructions (#10) and contribution instructions (#12)
