+++ share/wgs2ncbi.ini
@@ -24,13 +24,13 @@ authority=gnl|NaturalisBC|
 ; whose records can be scaffolds, contigs, chromosomes. The names of these sequences must 
 ; match those in the annotations.
 datadir=share/tblfasta
-datafile=share/king_cobra_scaffolds_spring_2011-cleaned.fasta
+datafile=share/example.fasta
 
 ; Annotations are in GFF3 format. The gff3file contains all the annotations in a single 
 ; file, whose contents are exploded into sequence-specific annotations as separate files 
 ; in gff3dir
 gff3dir=share/gff3
-gff3file=share/cobra.functional-cleaned.gff
+gff3file=share/example.gff
 
 ; Where to write *.sqn and validation files.
 outdir=share/asn1val
