### 2019-04-11 #1364 @rvosa
on it, as I hope you can see :)

On Thu, Apr 11, 2019 at 2:37 PM Andreas Sjödin <notifications@github.com>
wrote:

> @rvosa <https://github.com/rvosa> I started to add issues to the wgs2ncbi
> repo regarding failing automatic tests, stalled help command and
> contribution documentation. Could you please take a look at them?
>
