### 2019-04-15 6cb056c
added Usage section, including expected output for each step (#13, #14)
