+++ README.md
@@ -2,162 +2,245 @@ WGS2NCBI - toolkit for preparing genomes for submission to NCBI
 ===============================================================
 
 The process of going from an annotated genome to a valid NCBI submission is somewhat 
-cumbersome. "Boutique" genome projects might produce a scaffolded assembly in FASTA format
-and predicted genes in GFF3 tabular format (e.g. produced by the "maker" pipeline) but no 
-convenient tools appear to exist to turn these results in the format that NCBI requires.
-This project remedies this by providing some Perl scripts (with no dependencies except 
-URI::Escape) to do the re-formatting. Included is also a shell script that chains the Perl 
-scripts together and runs NCBI's tbl2asn on the result. This shell script is intended as 
-an example and should be edited or copied to provide the right values.
-
-Intro: What needs to be submitted
----------------------------------
+cumbersome. "Boutique" genome projects typically produce a scaffolded assembly in FASTA 
+format (produced by any of a variety of de-novo assemblers) and predicted genes in GFF3 
+tabular format (e.g. produced by the "maker" pipeline) but no convenient tools appear to 
+exist to turn these results in the format that NCBI accepts.
 
 NCBI requires that "whole genome shotgunning" (WGS) genomes are submitted as .sqn files
 (under the new rules of Spring 2013, one file for each scaffold). A sqn file is a file 
 in ASN.1 syntax that contains both the sequence, its features, and the metadata about the 
 submission (i.e. the authors, the publication title, the organism, etc.). sqn files are 
-normally produced by the SeqIn program, which has a graphical user interface and which is 
-therefore not practical for the potentially many thousands of files that comprise an 
-entire genome. It is therefore preferred to use the program tbl2asn (command line), which 
-takes a directory with FASTA files (.fsa) and corresponding files with the gene 
-features in tabular format (.tbl), and a submission template (template.sbt) to produce
-the sqn files. We therefore need to do some data processing to prepare the inputs for 
-tbl2asn.
+normally produced by the SeqIn program, which has a graphical user interface. SeqIn works
+fine for a single gene or for a small genome (e.g. a mitochondrial genome) but for large
+genomes with thousands of genes spread out over potentially thousands of scaffolds the
+submission process done in this way is unworkable.
 
-Since our starting material is one giant FASTA file that contains the scaffolds and a GFF 
-file with the features, we need to do the following:
+The alternative is to use [tbl2asn](https://www.ncbi.nlm.nih.gov/genbank/tbl2asn2/) 
+command line program, which takes a directory with FASTA files (.fsa) and corresponding 
+files with the gene features in tabular format (.tbl), and a submission template 
+(template.sbt) to produce the sqn files. The trick thus becomes to convert the assembly
+FASTA file and the annotation GFF3 file into a collection of FASTA chunks with
+corresponding feature tables. This is doable in principle (several toolkits provide
+generic convertors), but NCBI places quite a few restrictions on what are permissible 
+things to have in the FASTA headers, what coordinate ranges are credible as gene features,
+and what gene and gene product names are acceptable.
 
-1. use the GenBank web form to 
-   [create the submission template](#creating-the-submission-template)
-2. [explode the GFF3 file](#splitting-the-gff3-file) into smaller ones, one for each 
-   scaffold: `wgs2ncbi prepare -conf <config.ini>`
-3. [split the FASTA file](#splitting-the-fasta-file) into scaffolds and feature tables: 
-   `wgs2ncbi process -conf <config.ini>`
-4. [run tbl2asn](#running-tbl2asn) on the folder with the intermediate files: 
-   `wgs2ncbi convert -conf <config.ini>`
[…]
