### 2019-04-15 d699b93
minor README re-formatting
