+++ README.md
@@ -4,21 +4,23 @@ WGS2NCBI - preparing genomes for submission to NCBI
 The process of going from an annotated genome to a valid NCBI submission is somewhat 
 cumbersome. "Boutique" genome projects typically produce a scaffolded assembly in FASTA 
 format as produced by any of a variety of de-novo assemblers and predicted genes in GFF3 
-tabular format, e.g. as produced by the "maker" pipeline, but no convenient tools appear 
-to exist to turn these results in the format that NCBI accepts.
+tabular format, e.g. as produced by the 
+[maker](http://www.yandell-lab.org/software/maker.html) pipeline, but no convenient tools 
+appear to exist to turn these results in a format and to a standard that NCBI accepts.
 
-NCBI requires that "whole genome shotgunning" (WGS) genomes are submitted as .sqn files. 
+NCBI requires that "whole genome shotgunning" (WGS) genomes are submitted as `.sqn` files. 
 A sqn file is a file in ASN.1 syntax that contains both the sequence, its features, and 
 the metadata about the submission, i.e. the authors, the publication title, the organism, 
-etc.. sqn files are normally produced by the SeqIn program, which has a graphical user 
-interface. SeqIn works fine for a single gene or for a small genome (e.g. a mitochondrial 
+etc.. `.sqn` files are normally produced by the 
+[sequin](https://www.ncbi.nlm.nih.gov/Sequin/) program, which has a graphical user 
+interface. Sequin works fine for a single gene or for a small genome (e.g. a mitochondrial 
 genome) but for large genomes with thousands of genes spread out over potentially 
-thousands of scaffolds the submission process done in this way is unworkable.
+thousands of scaffolds the submission process done in this way would be unworkable.
 
-The alternative is to use [tbl2asn](https://www.ncbi.nlm.nih.gov/genbank/tbl2asn2/) 
-command line program, which takes a directory with FASTA files (.fsa) and corresponding 
-files with the gene features in tabular format (.tbl), and a submission template 
-(template.sbt) to produce the sqn files. The trick thus becomes to convert the assembly
+The alternative is to use the [tbl2asn](https://www.ncbi.nlm.nih.gov/genbank/tbl2asn2/) 
+command line program, which takes a directory with FASTA files (`.fsa`) and corresponding 
+files with the gene features in tabular format (`.tbl`), and a submission template 
+(template.sbt) to produce the `.sqn` files. The trick thus becomes to convert the assembly
 FASTA file and the annotation GFF3 file into a collection of FASTA chunks with
 corresponding feature tables. This is doable in principle - several toolkits provide
 generic convertors -, but NCBI places quite a few restrictions on what are permissible 
@@ -88,7 +90,7 @@ Before issuing any commands, the following steps need to be taken:
 3. You will need to prepare a 
    [submission template](http://www.ncbi.nlm.nih.gov/WebSub/template.cgi). The file
    [template.sbt](share/template.sbt) is an example of what these files look like.
-4. You need to have created a number of INI files correctly. Using the linked files as 
+4. You need to have created a number of `.ini` files correctly. Using the linked files as 
    examples, the following need to be prepared:
    - [wgs2ncbi.ini](share/wgs2ncbi.ini) - the main configuration file, in which you 
      specify the locations of the input files and output directories. In addition, here
@@ -128,7 +130,7 @@ also need to provide the location of this same file.
 
 What happens during this step is that the GFF3 file is pre-processed so that the following
 steps will have quicker access to the relevant contents than they would have if they had
-the scan through the entire file every time. To be precise, the following happens:
[…]
