### 2019-04-17 #1364 @druvus
thanks @rvosa ! I guess you should release an updated tagged version and update the manuscript to point to this new improved version.
