### 2019-04-17 #1364 @druvus
@lpantano I happy with wgs2ncbi once the article is updated to include a new updated version based on latest wgs2ncbi commit.
