### 2019-04-17 #1364 @rvosa
I tagged the most recent commit as v1.0.1 but I forgot where I need to insert that tag in the submission?
