### 2019-04-18 #1364 @rvosa
> @rvosa <https://github.com/rvosa>, can you create a Zenodo archive of
> your repository and give me the DOI? The title needs to match the papers
> title and make sure authors are correct in Zenodo as well.
>

Tagged release v1.0.3 is now mirrored on Zenodo, with the same title as the
manuscript, and with the coauthors embedded in the .zenodo.json file. The
DOI for this release is  10.5281/zenodo.2645762

>
