+++ README.md
@@ -1,3 +1,4 @@
+[![Build Status](https://travis-ci.org/naturalis/wgs2ncbi.svg?branch=master)](https://travis-ci.org/naturalis/wgs2ncbi)
 [![DOI](https://zenodo.org/badge/11924393.svg)](https://zenodo.org/badge/latestdoi/11924393)
 [![status](http://joss.theoj.org/papers/72b726934646d999001e9997ad2fdb54/status.svg)](http://joss.theoj.org/papers/72b726934646d999001e9997ad2fdb54)
 
