### 2019-04-19 #1364 @lpantano
Ok, @openjournals/joss-eics I think we are done. Please let me know if something is off.
